// src/model/config/enums/SubredditSourceOptionsEnum.ts
enum SubredditSourceOptionsEnum {
  SubscribedSubreddits = "Subscribed Subreddits",
  RedditUsersOnly = "Reddit Users Only",
  FrontPage = "Front Page",
  RedditListDotComRecentActivity = "RedditList.com Recent Activity",
  RedditListDotComSubscribers = "RedditList.com Subscribers",
  RedditListDotCom24HourGrowth = "RedditList.com 24 Hour Growth",
  SelectedSubRedditLists = "Selected Sub Reddit Lists",
  AllSubRedditLists = "All Sub Reddit Lists",
}

export default SubredditSourceOptionsEnum;